# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Lesson service",
        "title": "Lesson_sel",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1",
    "basePath": "/",
    "paths": {
        "/Lesson/{id}": {
            "put": {
                "description": "update lesson",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lesson"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateLessonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lesson": {
            "get": {
                "description": "get all lessons",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lesson"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetLessonsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lesson/complete/{id}": {
            "put": {
                "description": "complete or reject lesson",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lesson"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CompleteOrRejectLessonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lesson/form/{id}": {
            "put": {
                "description": "form lesson",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lesson"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateLessonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/lesson/{id}": {
            "get": {
                "description": "get lesson by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lesson"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetLessonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "delete lesson",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "call"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/tl/count/{id}": {
            "put": {
                "description": "update dc count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DC"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "lesson id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tl/delete/{id}": {
            "delete": {
                "description": "delete dc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DC"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "lesson id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.LessonRequest": {
            "type": "object",
            "properties": {
                "Lesson_time": {
                    "type": "integer"
                },
                "date_accepted": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_formed": {
                    "type": "string"
                },
                "delivery_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lesson_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.CompleteOrRejectLessonResponse": {
            "type": "object",
            "properties": {
                "lessson_request": {
                    "description": "заявка на доставку",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ds.LessonRequest"
                        }
                    ]
                },
                "total_count": {
                    "description": "общее количество",
                    "type": "integer"
                }
            }
        },
...
        "models.GetLessonResponse": {
            "type": "object",
            "properties": {
                "lesson_request": {
                    "description": "заявка на доставку",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ds.LessonRequest"
                        }
                    ]
                },
                "task_items": {
                    "description": "карточки доставки",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TaskItemWithCount"
                    }
                },
                "tasks_count": {
                    "description": "общее число доставок",
                    "type": "integer"
                }
            }
        },
        "models.GetLessonsResponse": {
            "type": "object",
            "properties": {
                "lessons": {
                    "description": "список звонков",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.LessonRequest"
                    }
                }
            }
        },
...
        "models.TaskItemWithCount": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "forced": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "minutes": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.UpdateLessonResponse": {
            "type": "object",
            "properties": {
                "lesson_request": {
                    "description": "заявка на доставку",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ds.LessonRequest"
                        }
                    ]
                }
            }
        },
...
    }
}